requirements.txt
google-genai==1.32.0
fastapi==0.116.1
pymongo==4.14.1
python-dotenv==1.1.1
uvicorn==0.35.0
tqdm==4.67.1